+++ lib/main.dart
@@ -1,80 +1,13 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:signals/signals_flutter.dart';
 import 'package:hotkey_manager/hotkey_manager.dart';
 import 'package:screen_capturer/screen_capturer.dart';
-import 'package:image/image.dart' as img;
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:pasteboard/pasteboard.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:flex_color_picker/flex_color_picker.dart';
 import 'l10n/l10n.dart';
+import 'image.dart';
 import 'settings.dart';
-
-extension ColorX on Color {
-  img.ColorRgba8 toRgba8() {
-    return img.ColorRgba8(red, green, blue, alpha);
-  }
-}
-
-extension Uint8ListX on Uint8List {
-  /// Adds a text overlay to an image.
-  ///
-  /// This function takes a string of text and overlays it onto an image at the specified
-  /// coordinates. The text is drawn with a semi-transparent background to ensure readability.
-  ///
-  /// Parameters:
-  /// - `text` (String): The text to be added to the image.
-  /// - `x` (int, optional): The x-coordinate where the text will be placed. Default is 10.
-  /// - `y` (int, optional): The y-coordinate where the text will be placed. Default is 10.
-  /// - `textHeight` (int, optional): The height of the text background. Default is 30.
-  ///
-  /// Returns:
-  /// - `Uint8List`: The modified image with the text overlay. If the original image cannot be
-  ///   decoded, the function returns the original image.
-  ///
-  /// Example:
-  /// ```dart
-  /// Uint8List modifiedImage = addTextToImage('Hello, World!', x: 50, y: 50, textHeight: 40);
-  /// ```
-  Uint8List addTextToImage(
-    String text, {
-    int x = 10,
-    int y = 10,
-    int textHeight = 30,
-    Color textColor = Colors.black,
-    Color backgroundColor = Colors.white,
-  }) {
-    img.Image? originalImage = img.decodeImage(this);
-    if (originalImage == null) return this;
-
-    int maxWidth = originalImage.width - 20;
-
-    // Draw background
-    img.fillRect(
-      originalImage,
-      x1: x,
-      y1: y,
-      x2: x + maxWidth,
-      y2: y + textHeight,
-      color: backgroundColor.toRgba8(),
-    );
-
-    // Draw the timestamp on the image
-    img.drawString(
-      originalImage,
-      text,
-      x: 10,
-      y: 10,
-      font: img.arial24,
-      color: textColor.toRgba8(),
-    );
-
-    return Uint8List.fromList(img.encodePng(originalImage));
-  }
-}
 
 enum OverlayMenuTile {
   overlayTextColor,
@@ -338,18 +271,10 @@ class _MyHomePageState extends State<MyHomePage> {
           backgroundColor: settings.value.backgroundOverlayColor.value,
         );
 
-        await copyToClipboard(overlayImage);
+        await overlayImage.copyToClipboard();
         return overlayImage;
       });
     });
-  }
-
-  Future<void> copyToClipboard(Uint8List data) async {
-    final directory = await getTemporaryDirectory();
-    final file = File('${directory.path}/screenshot.png');
-    await file.writeAsBytes(data);
-
-    await Pasteboard.writeFiles([file.path]);
   }
 
   Future<Color> _showColorPicker(Color color) {
